+++ test/Microsoft.Dnx.ApplicationHost.FunctionalTests/ApplicationHostTestFixture.cs
@@ -1,13 +1,11 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.Dnx.CommonTestUtils;
-using Xunit;
+using Microsoft.Dnx.Testing;
 
 namespace Microsoft.Dnx.ApplicationHost.FunctionalTests
 {
-    [CollectionDefinition(nameof(ApplicationHostTestCollection))]
-    public class ApplicationHostTestCollection: ICollectionFixture<DnxRuntimeFixture>
+    public class ApplicationHostTestFixture : DnxSdkFunctionalTestBase
     {
     }
 }
