+++ test/Microsoft.Dnx.DesignTimeHost.FunctionalTests/DthStartupTests.cs
@@ -78,7 +78,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(DnxSdks))]
         public void DthStartup_GetProjectInformation(DnxSdk sdk)
         {
@@ -111,7 +111,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(ProtocolNegotiationTestData))]
         public void DthStartup_ProtocolNegotiation(DnxSdk sdk, int requestVersion, int expectVersion)
         {
@@ -127,7 +127,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(DnxSdks))]
         public void DthStartup_ProtocolNegotiation_ZeroIsNoAllowed(DnxSdk sdk)
         {
@@ -143,7 +143,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(RuntimeComponentsWithBothVersions))]
         public void DthCompilation_GetDiagnostics_OnEmptyConsoleApp(DnxSdk sdk, int protocolVersion)
         {
@@ -173,7 +173,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(RuntimeComponentsWithBothVersions))]
         public void DthCompilation_RestoreComplete_OnEmptyLibrary(DnxSdk sdk, int protocolVersion)
         {
@@ -206,7 +206,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(UnresolvedDependencyTestData))]
         public void DthCompilation_Initialize_UnresolvedDependency(DnxSdk sdk, int protocolVersion, string referenceType, string testProjectName,
                                                                    string expectedUnresolvedDependency, string expectedUnresolvedType)
@@ -267,7 +267,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(DnxSdks))]
         public void DthNegative_BrokenProjectPathInLockFile_V1(DnxSdk sdk)
         {
@@ -301,7 +301,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(DnxSdks))]
         public void DthNegative_BrokenProjectPathInLockFile_V2(DnxSdk sdk)
         {
@@ -338,7 +338,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(DnxSdks))]
         public void DthDependencies_UpdateGlobalJson_RefreshDependencies(DnxSdk sdk)
         {
@@ -399,7 +399,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(DnxSdks))]
         public void CompileModuleWithDeps(DnxSdk sdk)
         {
@@ -431,7 +431,7 @@ namespace Microsoft.Dnx.DesignTimeHost.FunctionalTests
             }
         }
 
-        [Theory]
+        [Theory, TraceTest]
         [MemberData(nameof(DnxSdks))]
         public void AddDepsReturnsReferences(DnxSdk sdk)
         {
